DEVELOPER-4007 fixed the templating problem
I found a simple solution to the drupal location problem! Very happy the solution was simple and easy to follow. Even better, I don't need to worry about injecting the location into the POJOs somehow! I'm basically templating the JSON and doing a replaceAll afterward which will work fine as I need to send a string across the wire anyway.